# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quotation Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  name?: string | null;
  role: string;
  createdAt: string;
  updatedAt: string;
}

export interface Client {
  id: string;
  name: string;
  company?: string | null;
  email: string;
  phone?: string | null;
  addressLine1?: string | null;
  addressLine2?: string | null;
  city?: string | null;
  region?: string | null;
  postalCode?: string | null;
  country?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ClientInput {
  name: string;
  company?: string | null;
  email: string;
  phone?: string | null;
  addressLine1?: string | null;
  addressLine2?: string | null;
  city?: string | null;
  region?: string | null;
  postalCode?: string | null;
  country?: string | null;
  notes?: string | null;
}

export interface LineItem {
  id: string;
  quotationId: string;
  description: string;
  quantity: string;
  unit: string;
  unitPrice: string;
  lineTotal: string;
  position: number;
}

export type LineItemInputUnit =
  (typeof LineItemInputUnit)[keyof typeof LineItemInputUnit];

export const LineItemInputUnit = {
  hours: "hours",
  days: "days",
  fixed: "fixed",
} as const;

export interface LineItemInput {
  description: string;
  quantity: number;
  unit?: LineItemInputUnit;
  unitPrice: number;
  position?: number;
}

export type QuotationInputDiscountType =
  | (typeof QuotationInputDiscountType)[keyof typeof QuotationInputDiscountType]
  | null;

export const QuotationInputDiscountType = {
  PERCENTAGE: "PERCENTAGE",
  FIXED: "FIXED",
} as const;

export type QuotationInputTemplate =
  (typeof QuotationInputTemplate)[keyof typeof QuotationInputTemplate];

export const QuotationInputTemplate = {
  MODERN: "MODERN",
  CLASSIC: "CLASSIC",
} as const;

export interface QuotationInput {
  clientId: string;
  issueDate: string;
  validUntil: string;
  currency?: string;
  discountType?: QuotationInputDiscountType;
  discountValue?: number;
  taxRate?: number;
  notes?: string | null;
  terms?: string | null;
  paymentUrl?: string | null;
  showQrCode?: boolean;
  template?: QuotationInputTemplate;
  lineItems: LineItemInput[];
}

export interface QuotationSummary {
  id: string;
  number: string;
  status: string;
  total: string;
  currency: string;
  issueDate: string;
  validUntil: string;
  createdAt?: string;
  clientId: string;
  clientName?: string | null;
  clientCompany?: string | null;
  template?: string;
  discountType?: string | null;
  discountValue?: string;
  taxRate?: string;
  subtotal?: string;
  discountAmount?: string;
  taxAmount?: string;
}

export type QuotationDetail = QuotationSummary & {
  notes?: string | null;
  terms?: string | null;
  paymentUrl?: string | null;
  showQrCode?: boolean;
  sentAt?: string | null;
  acceptedAt?: string | null;
  paidAt?: string | null;
  client?: Client;
  lineItems?: LineItem[];
};

export interface CompanySettings {
  id: string;
  name: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  region?: string | null;
  postalCode: string;
  country: string;
  phone?: string | null;
  email: string;
  website?: string | null;
  taxNumber?: string | null;
  logoUrl?: string | null;
  currency: string;
  defaultTaxRate: string;
  defaultTerms?: string | null;
  defaultNotes?: string | null;
  defaultTemplate: string;
  defaultPaymentUrl?: string | null;
  updatedAt: string;
}

export type SettingsInputDefaultTemplate =
  (typeof SettingsInputDefaultTemplate)[keyof typeof SettingsInputDefaultTemplate];

export const SettingsInputDefaultTemplate = {
  MODERN: "MODERN",
  CLASSIC: "CLASSIC",
} as const;

export interface SettingsInput {
  name: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  region?: string | null;
  postalCode: string;
  country: string;
  phone?: string | null;
  email: string;
  website?: string | null;
  taxNumber?: string | null;
  logoUrl?: string | null;
  currency: string;
  defaultTaxRate: number;
  defaultTerms?: string | null;
  defaultNotes?: string | null;
  defaultTemplate: SettingsInputDefaultTemplate;
  defaultPaymentUrl?: string | null;
}

export type DashboardDataStatusCounts = { [key: string]: number };

export interface DashboardData {
  statusCounts: DashboardDataStatusCounts;
  outstandingTotal: string;
  recentQuotations: QuotationSummary[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name?: string;
  size?: number;
  contentType?: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlResponseMetadata;
}

export type SeedUserBody = {
  email?: string;
  name?: string;
};

export type SeedUser200 = {
  user: User;
  seeded: boolean;
};

export type SyncUserBody = {
  email?: string;
  name?: string;
};

export type ListClientsParams = {
  search?: string;
};

export type ListQuotationsParams = {
  status?: string;
  clientId?: string;
};

export type ChangeQuotationStatusBodyStatus =
  (typeof ChangeQuotationStatusBodyStatus)[keyof typeof ChangeQuotationStatusBodyStatus];

export const ChangeQuotationStatusBodyStatus = {
  DRAFT: "DRAFT",
  SENT: "SENT",
  ACCEPTED: "ACCEPTED",
  REJECTED: "REJECTED",
  PAID: "PAID",
  EXPIRED: "EXPIRED",
} as const;

export type ChangeQuotationStatusBody = {
  status: ChangeQuotationStatusBodyStatus;
};

export type UploadLogoBody = {
  logo?: Blob;
};

export type UploadLogo200 = {
  url?: string;
  objectPath?: string;
};