# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quotation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/seed:
    post:
      operationId: seedUser
      tags: [auth]
      summary: First-run seed — upsert user and ensure company_settings row exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                name: { type: string }
      responses:
        "200":
          description: Seed result
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  seeded: { type: boolean }
                required: [user, seeded]
  /auth/sync:
    post:
      operationId: syncUser
      tags: [auth]
      summary: Sync Clerk user into DB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                name: { type: string }
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string, nullable: true }
        role: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, email, role, createdAt, updatedAt]
    Client:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string, nullable: true }
        email: { type: string }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, email, createdAt, updatedAt]
    ClientInput:
      type: object
      properties:
        name: { type: string }
        company: { type: string, nullable: true }
        email: { type: string }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [name, email]
...